pipeline/scripts/requirements.txt
requests
python-dotenv
sqlalchemy
psycopg[binary]
pgvector
numpy
sentence-transformers